+++ Framework-Headers/Frameworks/PhotosUI._framework/Objc/PXGEdgeStretchEffect.h
@@ -12,7 +12,7 @@
 
 - (id)initWithMetadataStore:(id)arg1;
 - (int)shaderFlags;
-- (int)type;
+- (unsigned short)type;
 
 @end
