+++ Framework-Headers/PrivateFrameworks/ChatKit._framework/Objc/CKDeviceOrientationManager.h
@@ -37,7 +37,7 @@ __attribute__((visibility("hidden")))
 - (void)_updateListeningState;
 - (BOOL)_wantsOrientationEvents;
 - (void)accelerometer:(id)arg1 didAccelerateWithTimeStamp:(double)arg2 x:(float)arg3 y:(float)arg4 z:(float)arg5 eventType:(int)arg6;
-- (void)accelerometer:(id)arg1 didChangeDeviceOrientation:(int)arg2;
+- (void)accelerometer:(id)arg1 didChangeDeviceOrientation:(long long)arg2;
 - (void)beginListeningForOrientationEventsWithKey:(id)arg1;
 - (void)dealloc;
 - (void)endListeningForOrientationEventsWithKey:(id)arg1;
