+++ Framework-Headers/PrivateFrameworks/CompanionSync._framework/Objc/SYStoreSessionOwner.h
@@ -36,7 +36,7 @@ __attribute__((visibility("hidden")))
 - (void).cxx_destruct;
 - (void)syncSession:(id)arg1 applyChanges:(id)arg2 completion:(CDUnknownBlockType)arg3;
 - (void)syncSession:(id)arg1 didEndWithError:(id)arg2;
-- (long long)syncSession:(id)arg1 enqueueChanges:(CDUnknownBlockType)arg2 error:(id *)arg3;
+- (unsigned int)syncSession:(id)arg1 enqueueChanges:(CDUnknownBlockType)arg2 error:(id *)arg3;
 - (BOOL)syncSession:(id)arg1 resetDataStoreWithError:(id *)arg2;
 - (BOOL)syncSession:(id)arg1 rollbackChangesWithError:(id *)arg2;
