+++ Framework-Headers/PrivateFrameworks/GeoServices._framework/Objc/GEOComposedWalkingUncertainArrivalRouteStep.h
@@ -11,15 +11,15 @@
 @interface GEOComposedWalkingUncertainArrivalRouteStep : GEOComposedWalkingRouteStep
 {
     NSArray *_aggregatedGEOSteps;
-    unsigned int _aggregatedStepsDistance;
+    double _aggregatedStepsDistance;
     unsigned int _aggregatedStepsDuration;
 }
 
 + (BOOL)supportsSecureCoding;
 - (void).cxx_destruct;
-- (unsigned int)_calculateAggregatedStepsDistance;
+- (double)_calculateAggregatedStepsDistance;
 - (unsigned int)_calculateAggregatedStepsDuration;
-- (unsigned int)distance;
+- (double)distance;
 - (unsigned int)duration;
 - (void)encodeWithCoder:(id)arg1;
 - (id)initWithCoder:(id)arg1;
